Re #142: Add support for loading full-res images
Corebird always loaded large size and shrunk it, which used lots
of memory. We now load "small" for the thumbnail and only load
full-res when the user requests it.

This includes some cool "lowsrc" trickery where we first show the user
the thumbnail scaled up while loading the high-res in the background.
On a fast Internet connection, this is barely distinguishable from
just showing the full-res image.